🎨🎛️🧰: move `InputLine` into `lively.components`
