+++ lively.components/list.js
@@ -1,8 +1,9 @@
-import { Morph, Text, Label, Icon, morph, touchInputDevice, part } from 'lively.morphic';
+import { Morph, Label, Icon, morph, touchInputDevice, part } from 'lively.morphic';
 import { pt, Color, Rectangle } from 'lively.graphics';
 import { arr, Path, string } from 'lively.lang';
 import { signal, noUpdate, once, connect } from 'lively.bindings';
 import { Button, ButtonModel } from './buttons.js';
+import { InputLine } from './inputs.js';
 
 export function asItem (obj) {
   // make sure that object is of the form
@@ -901,7 +902,7 @@ export class FilterableList extends Morph {
 
       submorphs: {
         initialize () {
-          const input = Text.makeInputLine({
+          const input = new InputLine({
             name: 'input',
             highlightWhenFocused: false,
             fixedHeight: false,
